Changes based on PR#2
